+++ feature_experimental_rpc/src/main/java/xyz/dead8309/feature_experimental_rpc/ExperimentalRpcScreen.kt
@@ -104,6 +104,10 @@ fun ExperimentalRpcScreen(
             .nestedScroll(scrollBehavior.nestedScrollConnection),
         topBar = {
             LargeTopAppBar(
+                colors = TopAppBarDefaults.largeTopAppBarColors(
+                    containerColor = MaterialTheme.colorScheme.surface,
+                    scrolledContainerColor = MaterialTheme.colorScheme.surface,
+                ),
                 title = {
                     Text(
                         text = stringResource(id = R.string.main_experimentalRpc),