+++ feature_media_rpc/src/main/java/com/my/axe/feature_media_rpc/MediaRpc.kt
@@ -113,6 +113,10 @@ fun MediaRPC(
             .nestedScroll(scrollBehavior.nestedScrollConnection),
         topBar = {
             LargeTopAppBar(
+                colors = TopAppBarDefaults.largeTopAppBarColors(
+                    containerColor = MaterialTheme.colorScheme.surface,
+                    scrolledContainerColor = MaterialTheme.colorScheme.surface,
+                ),
                 title = {
                     Text(
                         text = stringResource(id = R.string.main_mediaRpc),